refactor: 조회용 서비스의 분리
