+++ src/main/java/com/capstone/cschatbot/cs/service/query/CSQueryService.java
@@ -1,0 +1,11 @@
+package com.capstone.cschatbot.cs.service.query;
+
+import com.capstone.cschatbot.cs.dto.response.CSChatHistory;
+import com.capstone.cschatbot.cs.dto.response.CSChatHistoryList;
+
+public interface CSQueryService {
+
+    CSChatHistoryList findAllCSChat(String memberId);
+    CSChatHistoryList findAllCSChatByTopic(String memberId, String topic);
+    CSChatHistory findCSChat(String chatRoomId);
+}
